+++ frontend/src/components/ui/HomePageMap.tsx
@@ -1,14 +1,41 @@
 "use client";
+import React, {useState } from "react";
 import MarkerPin from "./marker";
 import Map from "./map";
+import NewTripButton from "./NewTripButton";
 
 function HomePageMap() {
+
+  const [isSelectedLocation, setSelected] = useState(false);
+  const [Location, setLocation] = useState({lat:43.553185075739556,lng:-79.42145117539064});
+
+  function MarkerFinishDrag(event: google.maps.MapMouseEvent) {
+    if (event.latLng) {
+        let lat = event.latLng.lat();
+        let lng = event.latLng.lng();
+        console.log("The final lat", lat);
+        console.log("The final lng", lng);
+        setSelected(true);
+        console.log(isSelectedLocation);
+        setLocation({ lat, lng });
+    }
+}
   
   return (
-    <Map>
-      <MarkerPin positionMarker={{lat:43.553185075739556,lng:-79.42145117539064} } 
-      draggable={true}/>
-    </Map>
+    <div className="relative w-full h-full">
+                <Map>
+                    <MarkerPin
+                        positionMarker={Location}
+                        draggable={true}
+                        onDragEnd={MarkerFinishDrag}
+                    />
+                </Map>
+                {isSelectedLocation && (
+                    <div className="absolute bottom-4 right-4">
+                        <NewTripButton position={Location} />
+                    </div>
+                )}
+        </div>
   )
 }
